% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{assignAllTables}
\alias{assignAllTables}
\title{Assign all tables from all servers to a symbol in DataSHIELD}
\usage{
assignAllTables(datasources)
}
\arguments{
\item{datasources}{A list of \link[DSI]{DSConnection-class} objects.}
}
\value{
A data.frame containing the table names.
}
\description{
The function returns the table names in case of success.
}
\examples{
conns <- DSI::datashield.connections_find()
tabNamesTidy <- assignAllTables(conns)
tabNamesTidy
#        id     .x
# 1 server1 table1
# 2 server2 table2
}
