% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{checkDatasources}
\alias{checkDatasources}
\title{Check datasources}
\usage{
checkDatasources(datasources = DSI::datashield.connections_find())
}
\arguments{
\item{datasources}{A list of \link[DSI]{DSConnection-class} objects.}
}
\value{
An integer. If datasources check passes, number of connections are returned, otherwise -1 is returned.
}
\description{
Checks if datasources are a list of DSConnection-class.
}
\examples{
checkDatasources()
# 2
}
