% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{createFactorVars}
\alias{createFactorVars}
\title{Create for each level of a (factor) variable a  boolean factor on the server
side in DataShield.}
\usage{
createFactorVars(x, datasources = DSI::datashield.connections_find())
}
\arguments{
\item{x}{A character string referring to a table column in DataShield.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}
}
\description{
Create for each level of a (factor) variable a  boolean factor on the server
side in DataShield.
}
\examples{
createFactorVars("tab1$GENDER", datasources=conns)
# [[1]]
# [[1]]$is.object.created
# [1] "A data object <GENDER0> has been created in all specified data sources"
#
# [[1]]$validity.check
# [1] "<GENDER0> appears valid in all sources"
#
#
# [[2]]
# [[2]]$is.object.created
# [1] "A data object <GENDER1> has been created in all specified data sources"
#
# [[2]]$validity.check
# [1] "<GENDER1> appears valid in all sources"
}
