% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds.boxplot_data}
\alias{ds.boxplot_data}
\title{Return boxplot data}
\usage{
ds.boxplot_data(
  x,
  variables = NULL,
  group = NULL,
  group2 = NULL,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{variables}{A character vector. Name of the requested columns of the DataShield table object.}

\item{group}{A character string or NULL. If ! NUL, then the boxplot data are for grouped boxplots for each group level of the variable "group".}

\item{group2}{A character string or NULL. If ! NUL, then the boxplot data are for grouped boxplots for each group level of the variable "group2".}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects.}

\item{tab}{A character string. Name of the DataShield table object.}
}
\value{
A data.frame containing the data type for each variable in table tab.
}
\description{
Returns the first quantile, median, third quantile and a data privacy preserving safe range (ymin and ymax).
}
\examples{
\dontrun{
  # connecting to the Opal servers
  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

  # Get the boxplot data
  boxplot_data <- ds.boxplot_data("D", "LAB_HDL", datasources=DSI::datashield.connections_find())
  print(boxplot_data$combined)
  #          x  ymin lower middle   upper    ymax
  #     <char> <num> <num>  <num>   <num>   <num>
  # 1: LAB_HDL 0.876 1.302  1.585 1.85025 2.22755

  # compute boxplot data, grouped according to GENDER and BMI CATEGORY
  boxplot_data <- ds.boxplot_data("D", "LAB_HDL", "GENDER", "PM_BMI_CATEGORICAL", datasources=DSI::datashield.connections_find())
  print(boxplot_data$combined)

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}

}
