% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds.meta}
\alias{ds.meta}
\title{Gets metadata for multiple variables of a table}
\usage{
ds.meta(x = NULL, datasources = NULL, simplify = TRUE)
}
\arguments{
\item{x}{a character string specifying the name of the table object.}

\item{datasources}{a list of \link[DSI:DSConnection-class]{DSConnection-class}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \link[DSI:datashield.connections_default]{datashield.connections_default()}.}

\item{simplify}{a boolean. If TRUE (default) the result is simplified to a data.frame. Otherwise a list is returned.}
}
\value{
\code{ds.metadata} returns to the metadata of the associated table held at the server.
}
\description{
This function gets the metadata for multiple variables of a
given object on the server. Basically the DataSHIELD function ds.metadata is
applied to all variables given in x.
By default the result is returned as a data.frame.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

  # Get the metadata associated with table 'D'
  metadata <- ds.meta(x = c('D$LAB_TSC','D$LAB_TRIG','D$LAB_HDL'), datasources = conns)
  print(metadata)
  #          label                     opal.value_type opal.entity_type opal.repeatable opal.index opal.nature
  # LAB_TSC  "Total Serum Cholesterol" "decimal"       "Participant"    0               0          "CONTINUOUS"
  # LAB_TRIG "Triglycerides"           "decimal"       "Participant"    0               0          "CONTINUOUS"
  # LAB_HDL  "HDL Cholesterol"         "decimal"       "Participant"    0               0          "CONTINUOUS"

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
