% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds_get_boxplot_data}
\alias{ds_get_boxplot_data}
\title{Get (ungrouped) boxplot data (combined) for all or some numeric variables of a table in unified form}
\usage{
ds_get_boxplot_data(
  tab,
  vars = NULL,
  message_fun = message,
  datasources = datashield.connections_find()
)
}
\arguments{
\item{tab}{A character string. Name of the DataShield table object.}

\item{vars}{A vector of character strings. Names of the variables in the table object `tab`.}

\item{datasources}{a list of \code{\link{dsBaseClient::DSConnection-class}}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \code{\link{dsBaseClient::datashield.connections_default}}.}
}
\value{
A data.frame containing the boxplot data for each variable in table tab.
}
\description{
The main purpose of this function is to be a helper function
for the function all_summaries.

It returns for each of the requested variables (`vars`) in table `tab` from the
`datasources` the boxplot data (for ungrouped boxplots only).
}
\examples{
ds_get_boxplot_data("D")
# A tibble: 1 × 3
 variable feature  value
 <chr>    <chr>    <chr>
 1 bmi_T1   missings 2228
}
