% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds_get_cat_summary}
\alias{ds_get_cat_summary}
\title{Get summary statistics for all or some categorical variables of a table in unified form}
\usage{
ds_get_cat_summary(
  tab,
  vars = NULL,
  check = T,
  shiny_notification = F,
  n_levels = NULL,
  datasources = datashield.connections_find()
)
}
\arguments{
\item{tab}{A character string. Name of the DataShield table object.}

\item{vars}{A vector of character strings. Names of the variables in the table object `tab`.}

\item{check}{Logical. If FALSE, then the check for nonnumerical variables is skipped. This will
lead to errors, if numeric variables are included in the parameter vars.}

\item{shiny_notification}{A logical or integer. If not `FALSE` the parameter specifies the number
of seconds that a notification is shown in case of errors.}

\item{n_levels}{Integer value. If known in advance, the number of possible level values can be given.
This can help to get reliable results in more cases.}

\item{datasources}{a list of \code{\link{dsBaseClient::DSConnection-class}}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \code{\link{dsBaseClient::datashield.connections_default}}.}
}
\value{
A data.frame containing the boxplot data for each variable in table tab.
}
\description{
The main purpose of this function is to be a helper function
for the function all_summaries.

It returns for each of the requested variables (`vars`) in table `tab` from the
`datasources` summary statistics such as the number of observations (total, valid and missings),
the levels (level labels, and number of observations per level) and the mode.
}
\examples{
ds_get_cat_summary("D")
# A tibble: 1 × 3
 variable feature  value
 <chr>    <chr>    <chr>
 1 bmi_T1   missings 2228
}
