% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds_get_hist}
\alias{ds_get_hist}
\title{Get histogram data (combined for all servers) for all or some numeric variables of a table in unified form}
\usage{
ds_get_hist(
  tab,
  vars = NULL,
  bins = 11,
  shiny_notification = F,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{tab}{A character string. Name of the DataShield table object.}

\item{vars}{A vector of character strings. Names of the variables in the table object `tab`.}

\item{bins}{A numeric. The number of bins for the histogram.}

\item{shiny_notification}{A logical or integer. If not `FALSE` the parameter specifies the number of seconds that a notification is shown in case of errors.}

\item{datasources}{a list of \link[DSI:DSConnection-class]{DSConnection-class}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \link[DSI:datashield.connections_default]{datashield.connections_default()}.}
}
\value{
A data.frame containing the histogram data for each variable in table tab.
}
\description{
The main purpose of this function is to be a helper function
for the function all_summaries.

It returns for each of the requested variables (`vars`) in table `tab` from the
`datasources` the features histograms information including bin positions and sizes.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

ds_get_hist("D", datasources=DSI::datashield.connections_find())
# A tibble: 6 × 3
#   variable feature value
#   <chr>    <chr>   <chr>
# 1 bmi_T1   breaks  [9.8445,11.2128,12.5812,13.9495,15.3179,16.6862,18.0546,19.4229,20.7912,22.1596,23…
# 2 bmi_T1   mids    [10.5287,11.897,13.2653,14.6337,16.002,17.3704,18.7387,20.1071,21.4754,22.8438,24.…
# 3 bmi_T1   counts  [0,7,113,410,434,267,166,88,69,54,23,22,10,9,4,3,0,2,0,0]
# 4 bmi_T1   density [0,0.0018,0.0289,0.1049,0.111,0.0683,0.0425,0.0225,0.0176,0.0138,0.0059,0.0056,0.0…
# 5 bmi_T1   min     11.0884345752723
# 6 bmi_T1   max     35.9674388820484

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
