% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds_get_quantile_mean}
\alias{ds_get_quantile_mean}
\title{Get quantiles and mean (combined or seperate for all servers) for all or some numeric variables of a table in unified form}
\usage{
ds_get_quantile_mean(
  tab,
  vars = NULL,
  combine = T,
  tidy = T,
  datasources = datashield.connections_find()
)
}
\arguments{
\item{tab}{A character string. Name of the DataShield table object.}

\item{vars}{A vector of character strings. Names of the variables in the table object `tab`.}

\item{combine}{Logical. If TRUE the results over all servers are combined. Otherwise results are returned for each server separately.}

\item{tidy}{Logical. If tidy (and combine not FALSE), then the result is formatted as a data.frame, otherwise results for each variable are returned as elements of a list.}

\item{datasources}{a list of \code{\link{dsBaseClient::DSConnection-class}}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \code{\link{dsBaseClient::datashield.connections_default}}.}
}
\value{
A data.frame containing the histogram data for each variable in table tab.
}
\description{
The main purpose of this function is to be a helper function
for the function all_summaries.

It returns for each of the requested variables (`vars`) in table `tab` from the
`datasources` the quantiles (5%, 10%, 25%, 50%, 75%, 90%, 95%) and the mean.
}
\examples{
ds_get_quantile_mean("D")
# A tibble: 1 × 3
 variable feature  value
 <chr>    <chr>    <chr>
 1 bmi_T1   missings 2228
}
