% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ds_quantileMean}
\alias{ds_quantileMean}
\title{Get quantiles and mean (separate or combined for all servers) for one numeric variable}
\usage{
ds_quantileMean(x, combine = T, datasources = datashield.connections_find())
}
\arguments{
\item{combine}{Logical. If TRUE the results over all servers are combined. Otherwise results are returned for each server separately.}

\item{datasources}{a list of \code{\link{dsBaseClient::DSConnection-class}}
objects obtained after login. If the \code{dsBaseClient::datasources} argument is not specified
the default set of connections will be used: see \code{\link{dsBaseClient::datashield.connections_default}}.}

\item{X}{A character string. The name of the DataShield object.}
}
\value{
A data.frame containing the histogram data for each variable in table tab.
}
\description{
The main purpose of this function is to be a helper function
for the function get_quantile_mean.

It returns for the requested variable (`X`) from the
`datasources` the quantiles (5%, 10%, 25%, 50%, 75%, 90%, 95%) and the mean.
}
\examples{
ds_quantileMean("D$LAB_TSC")
# A tibble: 1 × 3
 variable feature  value
 <chr>    <chr>    <chr>
 1 bmi_T1   missings 2228
}
