% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsBetter.metadata}
\alias{dsBetter.metadata}
\title{Returns the metadata for one (!) specified variable}
\usage{
dsBetter.metadata(
  x,
  datasources = DSI::datashield.connections_find(),
  silent = FALSE
)
}
\arguments{
\item{x}{a string character, specifying the variable}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}

\item{silent}{Logical. If TRUE, DataSHIELD errors are suppressed.}
}
\value{
a matrix containing the metadata.
}
\description{
This function returns metadata for a specified variable -
without unnecessary checks and outputs (as in the slow ds.metadata() from
dsBaseClient). Due to the restrictions in dsBase::metadataDS only the
following attributes are returned: 'names', 'spec', 'class', 'label',
'opal.value_type', 'opal.entity_type', 'opal.repeatable', 'opal.index',
'opal.nature'
The variable does not need to be specified in each server to obtain a result.
}
\details{
The function returns the metadata, obtained from attributes function.
}
\examples{
\dontrun{
  # connecting to the Opal servers
  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  # Then perform login in each server
  conns <- datashield.login(logins=logindata, assign = TRUE, symbol = "D")

  # Get the metadata associated with table 'D'
  dsBetter.metadata(x = 'D$LAB_TSC', datasources = conns)
  #           server                     label opal.value_type    opal.entity_type opal.repeatable opal.index   opal.nature  class       obj.name
  # LAB_TSC "study1" "Total Serum Cholesterol"       "decimal"       "Participant"               0          0  "CONTINUOUS"     NA    "D$LAB_TSC"
  # LAB_TSC "study3" "Total Serum Cholesterol"       "decimal"       "Participant"               0          0  "CONTINUOUS"     NA    "D$LAB_TSC"

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
