% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCallGLM}
\alias{dsCallGLM}
\title{Perform generalized linear model regression in a non-disclosive (k-nearest-neighbor) way.}
\usage{
dsCallGLM(
  formula = NULL,
  data = NULL,
  family = c("gaussian", "binomial", "poisson"),
  offset = NULL,
  weights = NULL,
  checks = FALSE,
  maxit = 20,
  CI = 0.95,
  viewIter = FALSE,
  viewVarCov = FALSE,
  viewCor = FALSE,
  datasources = NULL
)
}
\arguments{
\item{formula}{An object of class formula describing the model to be fitted.}

\item{data}{A character string specifying the name of an (optional) data frame that contains all of the variables in the GLM formula.}

\item{family}{A character string. Identifies the error distribution function to use in the model ("gaussian", "binomial" or "poisson").}

\item{offset}{A character string specifying the name of a variable to be used as an offset.}

\item{weights}{A character string specifying the name of a variable containing prior regression weights for the fitting process.}

\item{checks}{A logical. If TRUE ds.glm checks the structural integrity of the model. Default FALSE. For more information see Details.}

\item{maxit}{A numeric scalar denoting the maximum number of iterations that are permitted before non-convergence is declared.}

\item{CI}{A numeric value specifying the confidence interval (default 0.95).}

\item{viewIter}{A logical. If TRUE the results of the intermediate iterations are printed. If FALSE only final results are shown. Default FALSE.}

\item{viewVarCov}{A logical. If TRUE the variance-covariance matrix of parameter estimates is returned. Default FALSE.}

\item{viewCor}{A logical. If TRUE the correlation matrix of parameter estimates is returned. Default FALSE.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects.}
}
\value{
A list with parameters similar to glm objects.
}
\description{
The function basically calls ds.glm, but, the data table name is added to the output.
This function is basically a more comfortable wrapper for `dsBaseClient::ds.glm`.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

mod_height_ds_lm <- dsCallGLM(formula = "PM_BMI_CONTINUOUS ~ GENDER + LAB_TRIG",
                          data = "D",
                          family = "gaussian",
                          datasources = conns)

mod_height_ds_lm$coefficients
#               Estimate Std. Error   z-value      p-value  low0.95CI high0.95CI
# (Intercept) 26.2806538  0.1574577 166.90614 0.000000e+00 25.9720424 26.5892651
# GENDER1     -0.7579016  0.1521374  -4.98169 6.303141e-07 -1.0560855 -0.4597176
# LAB_TRIG     0.7469029  0.0481524  15.51123 2.912536e-54  0.6525259  0.8412798

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
