% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatAgg}
\alias{dsCatAgg}
\title{Summary for all categorical variables of a table across all servers}
\usage{
dsCatAgg(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string specifying the name of a data.frame.}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects.}
}
\value{
A data.frame.
}
\description{
Summary for all categorical variables of a table across all servers
}
\examples{
#dsCatAgg("tab1$GENDER")
#dsCatAgg("tab1$DIS_AMI")
}
