% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatSummary}
\alias{dsCatSummary}
\title{Summary for all non-numeric variables in a DataSHIELD table}
\usage{
dsCatSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string referring to a DataSHIELD table.}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects.}
}
\value{
A data.frame.
}
\description{
Summary for all non-numeric variables in a DataSHIELD table
}
\examples{
#dsCatSummary("tab1")
}
