% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatTidy}
\alias{dsCatTidy}
\title{A tidy summary for all non-numeric variables in a DataSHIELD table}
\usage{
dsCatTidy(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string referring to a DataSHIELD table.}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects.}
}
\value{
A data.frame.
}
\description{
A tidy summary for all non-numeric variables in a DataSHIELD table
}
\examples{
#dsCatTidy("tab1")
}
