% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatVarSummary}
\alias{dsCatVarSummary}
\title{Summary for categorical variable across all servers}
\usage{
dsCatVarSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{Name of referring to a data.frame column as string.}

\item{datasources}{A list of 'OpalConnection'.}
}
\value{
A data.frame.
}
\description{
Summary for categorical variable across all servers
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

dsCatVarSummary("D$GENDER")
#    server length level value
# 1 server1   2163     0  1092
# 2 server2   3088     0  1585
# 3 server1   2163     1  1071
# 4 server2   3088     1  1503

dsCatVarSummary("D$DIS_AMI")
#        id length   level value
# 1 invalid     NA invalid    NA


  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
