% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsGapply}
\alias{dsGapply}
\title{Apply a function to subsets of a table in DataSHIELD for each different outcome of a categorical grouping variable}
\usage{
dsGapply(
  x,
  G,
  FUN,
  lazy = FALSE,
  datasources = datashield.connections_find(),
  ...
)
}
\arguments{
\item{G}{A character string specifying the group variable in X according to which the data are split before applying function FUN.}

\item{FUN}{A function which creates a data.frame. It will be applied to groupwise subsets of the table X.}

\item{lazy}{A boolean. If TRUE then objects for groupwise tables are only created if they don't exist.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}

\item{...}{A additional parameters to be passed to FUN.}

\item{X}{A character string specifying a table which is available as a DataShield object.}
}
\value{
A data.frame containing the results of FUN, applied to subgroups G of table X.
}
\description{
This is useful if an analysis has to be performed for each factor level of a grouping variable.
}
\examples{
# get summary statistics for all variables grouped by GENDER
genderwise_summary <- dsGapply("tab1","GENDER",dsAnalysisTools::dsSummary, datasources=conns[1])
head(genderwise_summary)
#   GENDER variable feature   value       tabname
# 1      0  LAB_TSC       N 1092.00 tab1.GENDER.0
# 2      0  LAB_TSC      5\%    4.07 tab1.GENDER.0
# 3      0  LAB_TSC     10\%    4.50 tab1.GENDER.0
# 4      0  LAB_TSC     25\%    5.09 tab1.GENDER.0
# 5      0  LAB_TSC     50\%    5.90 tab1.GENDER.0
# 6      0  LAB_TSC     75\%    6.67 tab1.GENDER.0
}
