% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsGet_xyg}
\alias{dsGet_xyg}
\title{Get two numeric columns and a factor column of a DataShield data.frame in a non-disclosive (k-nearest-neighbor) way.}
\usage{
dsGet_xyg(
  x,
  y,
  g,
  tab,
  method = 1,
  k = 3,
  noise = 0.25,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{x}{A character string. Name of the x-variable in the DataShield table object.}

\item{y}{A character string. Name of the y-variable in the DataShield table object.}

\item{g}{A character string. Name of the factor variable (group) in the DataShield table object.}

\item{tab}{A character string. Name of the DataShield table object.}

\item{method}{method    A character string that specifies the method that is used to generated non-disclosive coordinates to be displayed in a scatter plot. This argument can be set as 'deterministic' (method=1) or 'probabilistic' (method=2). Default 'deteministic'.}

\item{k}{Numeric. Only used if 'deterministic' method is used. The number of the nearest neighbors for which their centroid is calculated. Default 3.}

\item{noise}{Numeric. Only used if 'probabilistic' method is used. The percentage of the initial variance that is used as the variance of the embedded noise if the argument method is set to 'probabilistic'.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects. Default is to use all findable connections.}
}
\value{
A data.frame with coloumns for x, y, group g and server (source).
}
\description{
The result can be used e.g. for a grouped scatterplot and for further analysis on the client side.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

pair_data <- dsGet_xyg("LAB_TSC","LAB_TRIG","GENDER","D")

head(pair_data)
#    LAB_TSC  LAB_TRIG GENDER  server
# 1 6.379269 1.4462051      0 server1
# 2 5.427365 4.3632667      0 server1
# 3 7.877391 3.2380028      0 server1
# 4 5.521844 2.1337910      0 server1
# 5 5.139187 2.5045780      0 server1
# 6 6.204872 0.6240902      0 server1

# a grouped scatterplot
plot(pair_data$LAB_TSC, pair_data$LAB_TRIG, col=as.numeric(pair_data$GENDER)+1)

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
