% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsGetPairs}
\alias{dsGetPairs}
\title{Get two numeric columns and a factor column of a DataShield data.frame in a non-disclosive (k-nearest-neighbor) way.}
\usage{
dsGetPairs(
  tab,
  vars,
  group,
  method = 1,
  k = 3,
  noise = 0.25,
  datasources = datashield.connections_find()
)
}
\arguments{
\item{tab}{A character string. Name of the DataShield table object.}

\item{vars}{A character vector. Names of the variables in the DataShield table.}

\item{group}{A character string. Name of the factor variable (group) in the DataShield table object.}

\item{method}{method    A character string that specifies the method that is used to generated non-disclosive coordinates to be displayed in a scatter plot. This argument can be set as 'deterministic' (method=1) or 'probabilistic' (method=2). Default 'deteministic'.}

\item{k}{Numeric. Only used if 'deterministic' method is used. The number of the nearest neighbors for which their centroid is calculated. Default 3.}

\item{noise}{Numeric. Only used if 'probabilistic' method is used. The percentage of the initial variance that is used as the variance of the embedded noise if the argument method is set to 'probabilistic'.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}
}
\value{
A data.frame with coloumns for x, y, group g and server (source).
}
\description{
This function is basically a wrapper for dsGet_xyg, with slightly different syntax and output.
}
\examples{
pair_data <- dsGetPairs("tab1",c("LAB_TSC","LAB_TRIG"),"GENDER")

head(pair_data)
#      xName    yName        x         y GENDER  server
#  1 LAB_TSC LAB_TRIG 6.379269 1.4462051      0 server1
#  2 LAB_TSC LAB_TRIG 5.427365 4.3632667      0 server1
#  3 LAB_TSC LAB_TRIG 7.877391 3.2380028      0 server1
#  4 LAB_TSC LAB_TRIG 5.521844 2.1337910      0 server1
#  5 LAB_TSC LAB_TRIG 5.139187 2.5045780      0 server1
#  6 LAB_TSC LAB_TRIG 6.204872 0.6240902      0 server1

# a grouped scatterplot
plot(pair_data$x, pair_data$y, col=as.numeric(pair_data$GENDER)+1)
}
