% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsGLM}
\alias{dsGLM}
\title{Perform generalized linear model regression and return the model object}
\usage{
dsGLM(
  formula = NULL,
  data = NULL,
  family = c("gaussian", "binomial", "poisson"),
  offset = NULL,
  weights = NULL,
  maxit = 20,
  CI = 0.95,
  viewIter = FALSE,
  viewCor = FALSE,
  datasources = NULL,
  add_tab_to_depvar = T
)
}
\arguments{
\item{formula}{An object of class formula describing the model to be fitted.}

\item{data}{A character string specifying the name of an (optional) data frame that contains all of the variables in the GLM formula.}

\item{family}{A character string. Identifies the error distribution function to use in the model ("gaussian", "binomial" or "poisson").}

\item{offset}{A character string specifying the name of a variable to be used as an offset.}

\item{weights}{A character string specifying the name of a variable containing prior regression weights for the fitting process.}

\item{maxit}{A numeric scalar denoting the maximum number of iterations that are permitted before non-convergence is declared.}

\item{CI}{A numeric value specifying the confidence interval (default 0.95).}

\item{viewIter}{A logical. If TRUE the results of the intermediate iterations are printed. If FALSE only final results are shown. Default FALSE.}

\item{viewCor}{A logical. If TRUE the correlation matrix of parameter estimates is returned. Default FALSE.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}

\item{add_tab_to_depvar}{A boolean. If TRUE, the table name + "$" is added as prefix to the dependent variable. Default is TRUE.}
}
\value{
A list with parameters similar to glm objects.
}
\description{
including non-disclosive k-nearest neighbor estimates for residuals.
}
\examples{
foo <- dsGLM(formula = "PM_BMI_CONTINUOUS ~ GENDER + LAB_TRIG",
             data = "tab1",
             family = "gaussian",
             datasources = conns)
## qq plot
qqnorm(foo$y)
qqline(foo$y)

foo <- dsGLM(formula = "PM_BMI_CONTINUOUS ~ LAB_HDL + LAB_TRIG",
             data = "tab1", family = "gaussian", datasources = conns)
# histogram of privacy preserving residuals
hist(x = foo$residual, type = "combine", breaks = 18, datasources = conns)
# compare that histogram with `dsBaseClient::ds.histogram`
dsBaseClient::ds.histogram(x = "mod_residuals",
            type = "combine",
            num.breaks = 19,
            datasources = conns)
# scatterplot
dsBaseClient::ds.scatterPlot(x="mod_predicted", y="mod_residuals", type="combine", datasources=conns)

# attention: it is still not fully compatible with the models from stats::lm:
# e.g. this does not work due to lack of the data slot:
#stats:::plot.lm(foo)
}
