% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsNumSummary}
\alias{dsNumSummary}
\title{A summary for a numeric variable in a DataSHIELD table}
\usage{
dsNumSummary(x, datasources = DSI::datashield.connections_find())
}
\arguments{
\item{x}{A character string referring to a numeric column in a DataSHIELD table.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects.
Default is to use all findable datasources.}
}
\value{
A data.frame.
}
\description{
The summary contains the number of observations, quantiles (5%, 10%, 25%, 50%,
75%, 90%, 95%) and the mean.
}
\details{
A similar function for all columns of a table is `dsNumVarSummary` which returns results for all numeric columns of a table -- however it does not
return the number of observations.
}
\examples{
dsNumSummary("tab1$LAB_HDL")
#    LAB_HDL        value
#  1       N 5251.0000000
#  2      5\%    0.8606589
#  3     10\%    1.0385205
#  4     25\%    1.2964949
#  5     50\%    1.5704848
#  6     75\%    1.8418712
#  7     90\%    2.0824057
#  8     95\%    2.2191369
#  9    Mean    1.5619572
}
