% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsSubsetLevels}
\alias{dsSubsetLevels}
\title{Get subsets of a DataShield object for each factor level of a factor variable}
\usage{
dsSubsetLevels(
  var,
  tab,
  lazy = FALSE,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{var}{A character string. Name of the variable in the DataShield table object.}

\item{tab}{A character string. Name of the DataShield table object.}

\item{lazy}{A boolean. If TRUE then objects for groupwise tables are only created if they don't exist.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable connections.}
}
\value{
A list of the names of the created DataShield objects.
}
\description{
Get subsets of a DataShield object for each factor level of a factor variable
}
\examples{
dsSubsetLevels("PM_BMI_CATEGORICAL","tab1")
# then these new tables canbe used, e.g. for a plot:
ggHistDS("tab1.PM_BMI_CATEGORICAL.1$LAB_TRIG")
}
