% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsSummary}
\alias{dsSummary}
\title{A summary for all (numeric and non-numeric) variables in a DataSHIELD table}
\usage{
dsSummary(x = NULL, datasources = DSI::datashield.connections_find())
}
\arguments{
\item{x}{A character string referring to a numeric column in a DataSHIELD table.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects.}
}
\value{
A data.frame.
}
\description{
A summary for all (numeric and non-numeric) variables in a DataSHIELD table
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

rs <- dsSummary("D")
head(rs, 20)
#              variable feature   value
# 1             LAB_TSC       N 5251.00
# 2             LAB_TSC      5\%    4.12
# 3             LAB_TSC     10\%    4.50
# 4             LAB_TSC     25\%    5.13
# 5             LAB_TSC     50\%    5.83
# 6             LAB_TSC     75\%    6.54
# 7             LAB_TSC     90\%    7.23
# 8             LAB_TSC     95\%    7.65
# 9             LAB_TSC    Mean    5.86
# 10           LAB_TRIG       N 5251.00
# 11           LAB_TRIG      5\%   -0.47
# 12           LAB_TRIG     10\%    0.08
# 13           LAB_TRIG     25\%    1.02
# 14           LAB_TRIG     50\%    2.08
# 15           LAB_TRIG     75\%    3.09
# 16           LAB_TRIG     90\%    4.05
# 17           LAB_TRIG     95\%    4.59
# 18           LAB_TRIG    Mean    2.07
# 19            LAB_HDL       N 5251.00
# 20            LAB_HDL      5\%    0.86

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
