% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsTableSummary}
\alias{dsTableSummary}
\title{Summary for a table in data shield}
\usage{
dsTableSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{Table name as string.}

\item{datasources}{A list of 'OpalConnection'.}
}
\value{
A data.frame
}
\description{
Output: variables per server
}
\details{
The output is more convenient than using the ds.summary function directly.
}
\examples{
#dsTableSummary("tab1")
}
