% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsUniqueVarnames}
\alias{dsUniqueVarnames}
\title{Get unique variable names of a table object across all servers}
\usage{
dsUniqueVarnames(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{Table name as string.}

\item{datasources}{A list of 'OpalConnection'.}
}
\value{
A vector.
}
\description{
Get unique variable names of a table object across all servers
}
\examples{
#dsUniqueVarnames("tab1")
}
