% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{get_xy}
\alias{get_xy}
\title{Get two numeric columns of a DataShield data.frame in a non-disclosive (k-nearest-neighbor) way.}
\usage{
get_xy(
  x,
  y,
  method = 1,
  k = 3,
  noise = 0.25,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{x}{A character string. Name of the x-variable in the DataShield table object.}

\item{y}{A character string. Name of the y-variable in the DataShield table object.}

\item{method}{method    A character string that specifies the method that is used to generated non-disclosive coordinates to be displayed in a scatter plot. This argument can be set as 'deterministic' (method=1) or 'probabilistic' (method=2). Default 'deteministic'.}

\item{k}{Numeric. Only used if 'deterministic' method is used. The number of the nearest neighbors for which their centroid is calculated. Default 3.}

\item{noise}{Numeric. Only used if 'probabilistic' method is used. The percentage of the initial variance that is used as the variance of the embedded noise if the argument method is set to 'probabilistic'.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects. Default is to use all findable connections.}
}
\value{
A data.frame with coloumns for x, y and server (source).
}
\description{
These data can than be used e.g. for a scatterplot or an analysis on the client side.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

pair_data <- get_xy("D$LAB_TSC","D$LAB_TRIG")

head(pair_data)
#          x         y  server
# 1 7.220168 2.4789300 server1
# 2 6.033267 1.3209085 server1
# 3 6.378069 1.4572257 server1
# 4 6.844664 4.4676767 server1
# 5 6.441833 0.7183122 server1
# 6 6.052160 2.9344455 server1

# a simple scatterplot
plot(pair_data$x, pair_data$y)

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
