% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{getRange}
\alias{getRange}
\title{Get Range}
\usage{
getRange(
  x,
  type = "combined",
  k = 3,
  noise = 0.25,
  safemode = F,
  datasources = DSI::datashield.connections_find(),
  stdnames = names(datasources),
  method = 1
)
}
\arguments{
\item{x}{A character string referring to a numeric column in a DataSHIELD table.}

\item{type}{Boolean. If "combined" then the results are aggregated over all servers,
otherwise a list of all server results is returned.}

\item{k}{Numeric. Only used if `method==2`. The number of the nearest neighbors for which their centroid is calculated. Default 3.}

\item{noise}{Numeric. Only used if `method==3`. The percentage of the initial variance that is used as the variance of the embedded noise if the argument method is set to 'probabilistic'.}

\item{safemode}{Boolean, if TRUE, histogramDS1 is called for each datasource
separately, to ensure that a result can be returned even if one datasource creates an error.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects. Default
is to use all findable datasources.}

\item{stdnames}{Character vector. If there are more than one datasources and
type is not "combined", then this defines how the rows of the resulting table
are labelled. Default is to use the names of the datasources.}

\item{method}{A numeric. Either 1, 2 or 3 indicating the method of disclosure
control that is used for the generation of the histogram. If the value is equal to 1 then the
'smallCellsRule' is used. If the value is equal to 2 then the 'deterministic' method is used.
If the value is set to 3 then the 'probabilistic' method is used.}
}
\value{
A numeric vector (if '"combined" or only 1 study) or a matrix (else).
}
\description{
Uses the histogramDS1 server function to get a safe range (spanning over
minimum and maximum, were the exact range is increased with some privacy
preserving noise).
}
\details{
If safemode is FALSE the function will fail if histogramDS1 fails for any
server.
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

getRange("D$LAB_TRIG")
#       min       max
# -3.483958 12.128934
getRange("D$LAB_TRIG", safemode=T)
#       min       max
# -3.483958 12.128934
getRange("D$LAB_TRIG", type="separate", safemode=T)
#               min      max
# server1 -3.483958 12.12893
# server2 -3.331776 12.02544

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
