% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{ggHistDS}
\alias{ggHistDS}
\title{Plot data in a histogram with R graphics}
\usage{
ggHistDS(
  x,
  bins = "Sturges",
  plot = "combined",
  k = 3,
  noise = 0.25,
  freq = T,
  safemode = F,
  range = NULL,
  datasources = DSI::datashield.connections_find(),
  stdnames = names(datasources),
  method = 2
)
}
\arguments{
\item{x}{A character string referring to a numeric column in a DataSHIELD table.}

\item{bins}{either a numeric defining the number of bins or "Sturges" to
choose the bin number automatically. Default is "Sturges".}

\item{plot}{If "combined" then a plot combining the results from all servers,
if "separate" plots for each datasource, if "all" plots for each datasource
and combined are plotted. Otherwise only data are returned as a list and no
plot is generated.}

\item{k}{Numeric. Only used if `method==2`. The number of the nearest neighbors for which their centroid is calculated. Default 3.}

\item{noise}{Numeric. Only used if `method==3`. The percentage of the initial variance that is used as the variance of the embedded noise if the argument method is set to 'probabilistic'.}

\item{freq}{Boolean.}

\item{safemode}{Boolean. If FALSE, no plot is generated if any datasource fails,
otherwise the failing datasources are ignored and the results shouw only
data from non-failing datasources.}

\item{range}{Numeric vector of length 2. A range for the histogram can be specified.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects.}

\item{stdnames}{Character vector. If there are more than one datasources and
type is not "combined", then this defines how the rows of the resulting table
are labelled. Default is to use the names of the datasources.}

\item{method}{A numeric. Either 1, 2 or 3 indicating the method of disclosure
control that is used for the generation of the histogram. If the value is equal to 1 then (only!) the
'smallCellsRule' is used. If the value is equal to 2 then the 'deterministic' method is used additionally to method 1 for bins with count zero according to method 1.
If the value is set to 3 then the 'probabilistic' method is used additionally to method 1 for bins with count zero according to method 1.}
}
\value{
A list containing the histogram object.
}
\description{
Plot data in a histogram with R graphics
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

ggHistDS("D$LAB_TRIG")
ggHistDS("D$LAB_TRIG", safemode=T, plot="separate")
ggHistDS("D$LAB_TRIG", safemode=T, plot="all")
ggHistDS("D$LAB_TRIG", safemode=T, plot=F)
# [[1]]
# [[1]]$histobject
# $breaks
# [1] -3.4839583 -2.2829666 -1.0819749  0.1190168  1.3200085  2.5210003  3.7219920  4.9229837  6.1239754  7.3249671  8.5259588  9.7269505
# [13] 10.9279422 12.1289340
#
# $counts
# [1]   4  33 135 396 531 450 192  50   6   1   0   0   3
#
# $density
# [1] 0.001849295 0.015256686 0.062413717 0.183080236 0.245493952 0.208045722 0.088766175 0.023116191 0.002773943 0.000000000 0.000000000
# [12] 0.000000000 0.001386971
#
# $mids
# [1] -2.8834624 -1.6824707 -0.4814790  0.7195127  1.9205044  3.1214961  4.3224878  5.5234795  6.7244713  7.9254630  9.1264547 10.3274464
# [13] 11.5284381
#
# $xname
# [1] "xvect"
#
# $equidist
# [1] TRUE
#
# attr(,"class")
# [1] "histogram"
#
# [[1]]$invalidcells
# [1] 1
#
# [[1]]$simulatedObs
# [1] 3
#
# [[1]]$uncertainCells
# [1] 10 11 12
#
# ....

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
