% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatAgg}
\alias{dsCatAgg}
\title{Summary for one (and only one!) categorical variable of a table across all servers}
\usage{
dsCatAgg(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string specifying the name of a data.frame.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects.}
}
\value{
A data.frame.
}
\description{
Summary for one (and only one!) categorical variable of a table across all servers
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

dsCatAgg("D$GENDER")
#   GENDER number of observations
# 1      0                   2677
# 2      1                   2574

dsCatAgg("D$DIS_AMI")
#   DIS_AMI number of observations
# 1 invalid                     NA

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
