% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatAgg}
\alias{dsCatAgg}
\title{Summary for one (and only one!) categorical variable of a table across all servers}
\usage{
dsCatAgg(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string specifying the name of a data.frame.}

\item{datasources}{A list of \code{\link{DSI::DSConnection-class}} objects.}
}
\value{
A data.frame.
}
\description{
Summary for one (and only one!) categorical variable of a table across all servers
}
\examples{
dsCatAgg("tab1$GENDER")
#   GENDER number of observations
# 1      0                   2677
# 2      1                   2574

dsCatAgg("tab1$DIS_AMI")
#   DIS_AMI number of observations
# 1 invalid                     NA
}
