% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatSummary}
\alias{dsCatSummary}
\title{Summary for all non-numeric variables in a DataSHIELD table}
\usage{
dsCatSummary(x = NULL, datasources = DSI::datashield.connections_find())
}
\arguments{
\item{x}{A character string referring to a DataSHIELD table.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects.}
}
\value{
A list of data.frame.
}
\description{
Summary for all non-numeric variables in a DataSHIELD table
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

dsCatSummary("D")
# $DIS_CVA
# level number of observations
# 1     0                   5248
# 2     1                      3
#
# $MEDI_LPD
# level number of observations
# 1     0                   5144
# 2     1                    107
#
# $DIS_DIAB
# level number of observations
# 1     0                   5174
# 2     1                     77
#
# $DIS_AMI
# level number of observations
# 1 invalid                     NA
#
# $GENDER
# level number of observations
# 1     0                   2677
# 2     1                   2574
#
# $PM_BMI_CATEGORICAL
# level number of observations
# 1     1                   1540
# 2     2                   1989
# 3     3                   1475

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
