% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatVarSummary}
\alias{dsCatVarSummary}
\title{Summary for categorical variable across all servers}
\usage{
dsCatVarSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{Name of referring to a data.frame column as string.}

\item{datasources}{A list of 'OpalConnection'.}
}
\value{
A data.frame.
}
\description{
Summary for categorical variable across all servers
}
\examples{
#dsCatVarSummary("tab1$GENDER")
#dsCatVarSummary("tab1$DIS_AMI")
}
