% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsCatVarSummary}
\alias{dsCatVarSummary}
\title{Summary for categorical variable across all servers}
\usage{
dsCatVarSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{Name of referring to a data.frame column as string.}

\item{datasources}{A list of 'OpalConnection'.}
}
\value{
A data.frame.
}
\description{
Summary for categorical variable across all servers
}
\examples{
dsCatVarSummary("tab1$GENDER")
#    server length level value
# 1 server1   2163     0  1092
# 2 server2   3088     0  1585
# 3 server1   2163     1  1071
# 4 server2   3088     1  1503

dsCatVarSummary("tab1$DIS_AMI")
#        id length   level value
# 1 invalid     NA invalid    NA
}
