% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsIsNumeric}
\alias{dsIsNumeric}
\title{Check if columns of a DataSHIELD table are numeric}
\usage{
dsIsNumeric(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string referring to a data.frame coloumn.}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects.}
}
\value{
A vector of boolean.
}
\description{
Check if columns of a DataSHIELD table are numeric
}
\examples{
#dsIsNumeric("tab1")
}
