% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsIsNumeric}
\alias{dsIsNumeric}
\title{Check if columns of a DataSHIELD table are numeric}
\usage{
dsIsNumeric(
  x = NULL,
  x.vars = NULL,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{x}{A character string referring to a data.frame.}

\item{x.vars}{Optional character string vector; a subset of the variables in table x, on which the function shall be applied}

\item{datasources}{A list of \code{\link{DSI::DSConnection-class}} objects. Default is to use all findable connections.}
}
\value{
A vector of boolean.
}
\description{
Check if columns of a DataSHIELD table are numeric
}
\examples{
dsIsNumeric("tab1")
# LAB_TSC           LAB_TRIG            LAB_HDL  LAB_GLUC_ADJUSTED  PM_BMI_CONTINUOUS
#    TRUE               TRUE               TRUE               TRUE               TRUE
# DIS_CVA           MEDI_LPD           DIS_DIAB            DIS_AMI             GENDER
#   FALSE              FALSE              FALSE              FALSE              FALSE
# PM_BMI_CATEGORICAL
#              FALSE
}
