% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsNumVarSummary}
\alias{dsNumVarSummary}
\title{Summary for all numeric variables in a DataSHIELD table}
\usage{
dsNumVarSummary(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{A character string referring to a DataSHIELD table.}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects.}
}
\value{
A data.frame.
}
\description{
Summary for all numeric variables in a DataSHIELD table
}
\examples{
#dsNumVarSummary("tab1")
}
