% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsSubsetLevels}
\alias{dsSubsetLevels}
\title{Get subsets of a DataShield object for each factor level of a factor variable}
\usage{
dsSubsetLevels(
  var,
  tab,
  lazy = FALSE,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{var}{A character string. Name of the variable in the DataShield table object.}

\item{tab}{A character string. Name of the DataShield table object.}

\item{lazy}{A boolean. If TRUE then objects for groupwise tables are only created if they don't exist.}

\item{datasources}{A list of \link[DSI:DSConnection-class]{DSConnection-class} objects. Default is to use all findable connections.}
}
\value{
A list of the names of the created DataShield objects.
}
\description{
Get subsets of a DataShield object for each factor level of a factor variable
}
\examples{
\dontrun{
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server="study1", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM1")
  builder$append(server="study3", url="https://opal-demo.obiba.org",
                 user="administrator", password="password",
                 table = "CNSIM.CNSIM3")
  logindata <- builder$build()

  conns <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

dsSubsetLevels("PM_BMI_CATEGORICAL","D")
# then these new tables canbe used, e.g. for a plot:
ggHistDS("D.PM_BMI_CATEGORICAL.1$LAB_TRIG")

  # clear the Datashield R sessions and logout
  DSI::datashield.logout(conns)
}
}
