% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsUniqueLevels}
\alias{dsUniqueLevels}
\title{Get factor levels for a variable over all server connections}
\usage{
dsUniqueLevels(
  var,
  tab = NULL,
  datasources = DSI::datashield.connections_find()
)
}
\arguments{
\item{var}{A character string. Name of the variable in the DataShield table object.}

\item{tab}{A character string. Name of the DataShield table object.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects. Default is to use all findable datasources.}
}
\value{
A numeric vector with the names of the factor levels.
}
\description{
Get factor levels for a variable over all server connections
}
\examples{
dsUniqueLevels("PM_BMI_CATEGORICAL","tab1")
# [1] "1" "2" "3"
dsUniqueLevels("tab1$PM_BMI_CATEGORICAL")
# [1] "1" "2" "3"
}
