% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{dsUniqueVarnames}
\alias{dsUniqueVarnames}
\title{Get unique variable names of a table object across all servers}
\usage{
dsUniqueVarnames(
  x = NULL,
  datasources = DSI::datashield.connections_find(),
  combined = T
)
}
\arguments{
\item{x}{Table name as string.}

\item{datasources}{A list of 'OpalConnection'. Default is to use all findable connections.}

\item{combined}{If FALSE, the available variables are returned per server, otherwise accross all servers.}
}
\value{
A vector.
}
\description{
Get unique variable names of a table object across all servers
}
\examples{
dsUniqueVarnames("tab1")
# output:
 [1] "LAB_TSC"            "LAB_TRIG"           "LAB_HDL"            "LAB_GLUC_ADJUSTED"  "PM_BMI_CONTINUOUS"  "DIS_CVA"            "MEDI_LPD"
 [8] "DIS_DIAB"           "DIS_AMI"            "GENDER"             "PM_BMI_CATEGORICAL"
}
