% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetools.R
\name{tabSummary}
\alias{tabSummary}
\title{Summarize the requested tables (DataSHIELD server list)}
\usage{
tabSummary(tabList, datasources)
}
\arguments{
\item{tabList}{A data.frame with the table names and server locations.}

\item{datasources}{A list of \link[DSI]{DSConnection-class} objects.}
}
\value{
A data.frame containing the table names, server location, symbols and dimensions.
}
\description{
Gets row and column counts for all given tables.
}
\examples{
conns <- DSI::datashield.connections_find()
tabSummary(tabNamesTidy, datasources=conns)
#     id  server   name rows cols
# 1 tab1 server1 table1 2163   11
# 2 tab2 server2 table2 3088   11
}
